@@ -5,33 +5,8 @@ import 'package:flutter_slidable/src/actions.dart';
 import 'package:flutter_slidable/src/slidable.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'common.dart';
-
 void main() {
   group('Slidable', () {
-    test('constructor asserts', () {
-      final values = [
-        true,
-        true,
-        Axis.horizontal,
-        DragStartBehavior.down,
-        true,
-        const SizedBox(),
-      ];
-
-      testConstructorAsserts(
-        values: values,
-        factory: (valueOrNull) => Slidable(
-          enabled: valueOrNull(0),
-          closeOnScroll: valueOrNull(1),
-          direction: valueOrNull(2),
-          dragStartBehavior: valueOrNull(3),
-          useTextDirection: valueOrNull(4),
-          child: valueOrNull(5),
-        ),
-      );
-    });
-
     testWidgets(
         'child should be able to open the horitzontal start action pane',
         (tester) async {
@@ -62,7 +37,7 @@ void main() {
               return GestureDetector(
                 key: gestureDetectorKey,
                 onTap: () {
-                  Slidable.of(context).openStartActionPane();
+                  Slidable.of(context)!.openStartActionPane();
                 },
               );
             }),
@@ -109,7 +84,7 @@ void main() {
               return GestureDetector(
                 key: gestureDetectorKey,
                 onTap: () {
-                  Slidable.of(context).openEndActionPane();
+                  Slidable.of(context)!.openEndActionPane();
                 },
               );
             }),
@@ -157,7 +132,7 @@ void main() {
               return GestureDetector(
                 key: gestureDetectorKey,
                 onTap: () {
-                  Slidable.of(context).openStartActionPane();
+                  Slidable.of(context)!.openStartActionPane();
                 },
               );
             }),
@@ -205,7 +180,7 @@ void main() {
               return GestureDetector(
                 key: gestureDetectorKey,
                 onTap: () {
-                  Slidable.of(context).openEndActionPane();
+                  Slidable.of(context)!.openEndActionPane();
                 },
               );
             }),
@@ -248,7 +223,7 @@ void main() {
               return GestureDetector(
                 key: gestureDetectorKey,
                 onTap: () {
-                  Slidable.of(context).openEndActionPane();
+                  Slidable.of(context)!.openEndActionPane();
                 },
               );
             },
@@ -288,7 +263,7 @@ void main() {
               return GestureDetector(
                 key: gestureDetectorKey,
                 onTap: () {
-                  Slidable.of(context).openEndActionPane();
+                  Slidable.of(context)!.openEndActionPane();
                 },
               );
             },
